revert test cases
